refactor common test utils
